+++ app/src/main/java/io/legado/app/ui/book/read/page/ContentTextView.kt
@@ -15,6 +15,7 @@ import io.legado.app.service.help.ReadBook
 import io.legado.app.ui.book.read.page.entities.*
 import io.legado.app.ui.book.read.page.provider.ChapterProvider
 import io.legado.app.ui.book.read.page.provider.ImageProvider
+import io.legado.app.ui.book.read.page.provider.TextPageFactory
 import io.legado.app.ui.widget.dialog.PhotoDialog
 import io.legado.app.utils.activity
 import io.legado.app.utils.getCompatColor
