+++ app/src/main/java/io/legado/app/ui/book/read/page/api/DataSource.kt
@@ -1,4 +1,4 @@
-package io.legado.app.ui.book.read.page
+package io.legado.app.ui.book.read.page.api
 
 import io.legado.app.service.help.ReadBook
 import io.legado.app.ui.book.read.page.entities.TextChapter
