+++ app/src/main/res/layout/activity_book_read.xml
@@ -4,8 +4,8 @@
     android:layout_height="match_parent"
     android:orientation="vertical">
 
-    <io.legado.app.ui.book.read.page.PageView
-        android:id="@+id/page_view"
+    <io.legado.app.ui.book.read.page.ReadView
+        android:id="@+id/read_view"
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
